Rename household_middleware file to middleware.py
Rename so that different types of middleware can go in this file
instead of those just for household views.

This change is made so that a middleware can be defined for ensuring
that a route can only be accessed by a user account whose email has
been confirmed.
