+++ app/views/household_shopping_list.py
@@ -4,7 +4,7 @@ from flask import Blueprint, jsonify, request
 from flask_login import current_user, login_required
 from models.household import Household, ShoppingListItem
 from models.user import User
-from app.utils.household_middleware import household_member_required
+from app.utils.middleware import household_member_required
 
 household_shopping_list_bl = Blueprint('household shopping list', __name__)
